ISO-3 country code list of beneficiaries for each ICS
resolves issue #11
